Cleaning the code. Getting ready for photon mapping

+++ src/main/java/com/ciebiada/reddot/SceneHandler.java
@@ -30,14 +30,14 @@ public class SceneHandler extends DefaultHandler {
         } else if (qName.equalsIgnoreCase("filter")) {
             String type = attributes.getValue("type");
 
-            float size = Float.valueOf(attributes.getValue("size"));
+            double size = Double.valueOf(attributes.getValue("size"));
 
             if (type.equalsIgnoreCase("gaussian")) {
-                float alpha = Float.valueOf(attributes.getValue("alpha"));
+                double alpha = Double.valueOf(attributes.getValue("alpha"));
                 filter = new Gaussian(size, alpha);
             } else if (type.equalsIgnoreCase("mitchell")) {
-                float b = Float.valueOf(attributes.getValue("b"));
-                float c = Float.valueOf(attributes.getValue("c"));
+                double b = Double.valueOf(attributes.getValue("b"));
+                double c = Double.valueOf(attributes.getValue("c"));
                 filter = new Mitchell(size, b, c);
             } else if (type.equalsIgnoreCase("tent")) {
                 filter = new Tent(size);
@@ -50,19 +50,19 @@ public class SceneHandler extends DefaultHandler {
             String type = attributes.getValue("type");
 
             String[] eyeStrs = attributes.getValue("eye").split(" ");
-            Vec eye = new Vec(Float.valueOf(eyeStrs[0]), Float.valueOf(eyeStrs[1]), Float.valueOf(eyeStrs[2]));
+            Vec eye = new Vec(Double.valueOf(eyeStrs[0]), Double.valueOf(eyeStrs[1]), Double.valueOf(eyeStrs[2]));
 
             String[] lookAtStrs = attributes.getValue("lookAt").split(" ");
-            Vec lookAt = new Vec(Float.valueOf(lookAtStrs[0]), Float.valueOf(lookAtStrs[1]), Float.valueOf(lookAtStrs[2]));
+            Vec lookAt = new Vec(Double.valueOf(lookAtStrs[0]), Double.valueOf(lookAtStrs[1]), Double.valueOf(lookAtStrs[2]));
 
-            float fov = Float.valueOf(attributes.getValue("fov"));
+            double fov = Double.valueOf(attributes.getValue("fov"));
 
             if (type.equalsIgnoreCase("pinhole")) {
-                scene.camera = new Pinhole(eye, lookAt.sub(eye).norm(), fov, (float) height / width);
+                scene.camera = new Pinhole(eye, lookAt.sub(eye).norm(), fov, (double) height / width);
             } else if (type.equalsIgnoreCase("thinLens")) {
-                float lensSize = Float.valueOf(attributes.getValue("lensSize"));
-                float distance = Float.valueOf(attributes.getValue("distance"));
-                scene.camera = new ThinLens(eye, lookAt.sub(eye).norm(), fov, (float) height / width, lensSize, distance);
+                double lensSize = Double.valueOf(attributes.getValue("lensSize"));
+                double distance = Double.valueOf(attributes.getValue("distance"));
+                scene.camera = new ThinLens(eye, lookAt.sub(eye).norm(), fov, (double) height / width, lensSize, distance);
             }
         } else if (qName.equalsIgnoreCase("reddot")) {
             Integer threads = Integer.valueOf(attributes.getValue("threads"));
